+++ src/model/qubo/cvrp/InfiniteCVRP.py
@@ -11,7 +11,7 @@ class InfiniteCVRP(ConstantCVRP):
         self,
         num_vehicles: int,
         distance_matrix: list[list[float]],
-        locations: list[tuple[int, int]],
+        locations: list[tuple[float, float]],
         simplify: bool,
         location_names: list[str] = None,
         distance_unit: DistanceUnit = DistanceUnit.METERS,
