+++ src/solver/qubo/DWaveSolver.py
@@ -48,7 +48,7 @@ class DWaveSolver(QuboSolver):
         self,
         num_vehicles: int,
         capacities: int | list[int] | None,
-        locations: list[tuple[int, int]],
+        locations: list[tuple[float, float]],
         trips: list[tuple[int, int, int]],
         use_rpp: bool,
         simplify=True,
